# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateExperimentRequest": {
        "additionalProperties": false,
        "properties": {
          "dataset_uri": {
            "description": "S3 URI of an approved dataset (s3://bucket/key).",
            "title": "Dataset Uri",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "file_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override the inferred format.",
            "title": "File Format"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "problem_type": {
            "$ref": "#/components/schemas/RequestedProblemType",
            "default": "auto"
          },
          "target_column": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Target Column",
            "type": "string"
          }
        },
        "required": [
          "name",
          "dataset_uri",
          "target_column"
        ],
        "title": "CreateExperimentRequest",
        "type": "object"
      },
      "CreateExperimentResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact_prefix": {
            "title": "Artifact Prefix",
            "type": "string"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentStatus"
          }
        },
        "required": [
          "experiment_id",
          "status",
          "artifact_prefix"
        ],
        "title": "CreateExperimentResponse",
        "type": "object"
      },
...
      "ExperimentListResponse": {
        "additionalProperties": false,
        "properties": {
          "available_roots": {
            "description": "Artifact buckets this platform may read.",
            "items": {
              "type": "string"
            },
            "title": "Available Roots",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ExperimentRecord"
            },
            "title": "Experiments",
            "type": "array"
          },
          "root": {
            "default": "",
            "description": "The artifact bucket these experiments were read from.",
            "title": "Root",
            "type": "string"
          }
        },
        "required": [
          "experiments",
          "count"
        ],
        "title": "ExperimentListResponse",
        "type": "object"
      },
...
      "FeatureReviewResponse": {
        "additionalProperties": false,
        "properties": {
          "decided": {
            "default": false,
            "description": "True once the user has saved a feature selection.",
            "title": "Decided",
            "type": "boolean"
          },
          "excluded_count": {
            "title": "Excluded Count",
            "type": "integer"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/FeatureReviewItem"
            },
            "title": "Features",
            "type": "array"
          },
          "problem_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemType"
              },
              {
                "type": "null"
              }
            ]
          },
          "selected_count": {
            "title": "Selected Count",
            "type": "integer"
          },
          "target_column": {
            "title": "Target Column",
            "type": "string"
          }
        },
        "required": [
          "experiment_id",
          "target_column",
          "problem_type",
          "features",
          "selected_count",
          "excluded_count"
        ],
        "title": "FeatureReviewResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nInternal platform that accelerates classical-ML POCs: automated EDA, deterministic data\nquality and leakage screening, supervised feature review, classical model training and\ncomparison, and \u2014 strictly after the deterministic results exist \u2014 semantic interpretation.\n",
    "title": "ML Factory",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/experiments": {
      "get": {
        "description": "Experiments found under an artifact bucket, newest first.",
        "operationId": "list_experiments_api_v1_experiments_get",
        "parameters": [
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Restrict to experiments you created.",
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "default": false,
              "description": "Restrict to experiments you created.",
              "title": "Mine",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Remote-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiments",
        "tags": [
          "experiments"
        ]
      },
      "post": {
        "description": "Create an experiment and start EDA. Returns immediately \u2014 AWS work runs asynchronously.",
        "operationId": "create_experiment_api_v1_experiments_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Remote-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExperimentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{experiment_id}": {
      "delete": {
        "description": "Soft delete: the record is hidden, the artifacts remain for audit.",
        "operationId": "delete_experiment_api_v1_experiments__experiment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Experiment",
        "tags": [
          "experiments"
        ]
      },
      "get": {
        "operationId": "get_experiment_api_v1_experiments__experiment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{experiment_id}/comparison": {
      "get": {
        "operationId": "get_comparison_api_v1_experiments__experiment_id__comparison_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Comparison",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/{experiment_id}/eda": {
      "get": {
        "operationId": "get_eda_api_v1_experiments__experiment_id__eda_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdaReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Eda",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/{experiment_id}/features": {
      "get": {
        "operationId": "get_features_api_v1_experiments__experiment_id__features_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Features",
        "tags": [
          "experiments"
        ]
      },
      "put": {
        "operationId": "put_features_api_v1_experiments__experiment_id__features_put",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Remote-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeatureSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureSelection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Features",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{experiment_id}/leakage": {
      "get": {
        "operationId": "get_leakage_api_v1_experiments__experiment_id__leakage_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeakageReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Leakage",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{experiment_id}/models/{model_name}": {
      "get": {
        "operationId": "get_model_api_v1_experiments__experiment_id__models__model_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/{experiment_id}/summary": {
      "get": {
        "operationId": "get_summary_api_v1_experiments__experiment_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Summary",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{experiment_id}/training": {
      "post": {
        "operationId": "start_training_api_v1_experiments__experiment_id__training_post",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Remote-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartTrainingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Training",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{experiment_id}/training-config": {
      "get": {
        "operationId": "get_training_config_api_v1_experiments__experiment_id__training_config_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Training Config",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{experiment_id}/training-status": {
      "get": {
        "operationId": "get_training_status_api_v1_experiments__experiment_id__training_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Training Status",
        "tags": [
          "experiments"
        ]
      }
    },
...
  }
}